Refine line widths at each breakpoint
The goal is to have a slightly shorter line measure than what
sphinx-basic-ng defaults to; we want about 70-80 characters per line.
Controlling the line width is fairly tricky because sphinx-basic-ng has
a lot of breakpoints. This css essentially implements each breakpoint
and refines the article container width.

Note that some adjustments will still need to be done for the "full
screen" view because we're not yet seeing/testing that.