Add webpack build for technote.css
This configuration is mostly based on Furo
https://github.com/pradyunsg/furo
This project doesn't use sphinx-theme-builder though.

The idea is that compileable sources go in ./assets. Now we can use Sass
to split the technote.css into multiple files, include normalize, and
also benefit for autoprefixer and minimization.

Similar for compiling JavaScript once we start adding that.

Since the built JS and CSS aren't committed, they need to be added to
MANIFEST.in for inclusion in the Python package (normally
setuptools_scm manages this).

Update path for technote.css. Webpack adds this file to static/styles
rather than static/ within the theme directory.