Add overall schema for a technote.toml config
Technotes are configured with a technote.toml file, in keeping with our
adoption of TOML for user-friendly configuration (and consistent with
the direction Python is going). Right now this toml configuration mostly
covers the basic bibliographic metadata needs for technote, though we'll
also need to expand the configuration scheme to handle technical
configuration for the Sphinx project.

Overall the technote.toml treats bibliographic metadata as optional.
This is keeping with the desire to make rich metadata something that you
can opt into and then the document will express that data to
end-users... but if you just want to start writing all you need to do is
start a document, write, and push to GitHub.