Get title from content first
The effect of this commit is to get the technote's title from the
content (first heading); if the technote.title field is set in
technote.toml that string is used instead.

To get the title from the technote, we're hooking into the
html-page-context event where the page has been parsed, but the Jinja
context for the HTML pages are still being set up.

We're storing the recovered page title in the TechnoteJinjaContext class
instance in the "technote" context variable; the title accessor returns
the appropriate title (title from content or title from technote.toml as
appropriate).