Add toc extension to customize toc structure
The default Sphinx local "toc" includes the page name. This extension
provides a new toc variable, technote_toc, that excludes this level
of the page hierarchy for cleaner styling.

This extension will also provide future means for adding a jump link to
the abstract, etc.

Finally, this commit improves the styling of the local toc (more
consistent indentation and spacing; borders to indicate hierarchy).