Handle technotes with no sections
When there are not sections in a technote, the original Sphinx toc won't
have a <ul> list below the level of the main title and therefore will be
empty since we extract that list of sections. The solution is to create
an empty <ul></ul> with beautifulsoup for for us to add the abstract, if
necessary.