Create an extension to override pygments css
The goal of this extension is to customize the Pygments style file so
that all pygments styles are selected with either a data-theme='light'
or 'dark' on the html element. This enables us to let the user switch
themes with javascript.

This approach to customizing the Pygments CSS is based on the work by
pydata-sphinx-theme.
https://github.com/pydata/pydata-sphinx-theme/blob/main/src/pydata_sphinx_theme/pygment.py

The light and dark themes are currently fixed. In the future we could
add a user setting for this. We're using the
github-(light|dark)-high-contrast themes from the accessible-pygments
package.