Make data-theme responsive to prefers-color-scheme
To make the data-theme selector for the pygments CSS responsive to the
user's color scheme preference, this commit adds JS that controls the
data-theme attribute. The theme can be light or dark. There is also a
data-mode attribute which can be light, dark, or auto (default). In
auto, the prefers-color-scheme media query is used.

Until we add UI for changing the color theme explicitly, this is the
default behaviour.

This is based on pydata-sphinx-theme;
https://github.com/pydata/pydata-sphinx-theme/blob/main/src/pydata_sphinx_theme/assets/scripts/pydata-sphinx-theme.js