Add CSS classes to hide elements for light/dark
Now that the light/dark theme is controlled via the data property on the
html element, we can't use media queries to choose the right image for a
given theme. Instead, introduce new CSS classes technote-themed-dark and
technote-themed-light that can be used to show that a specific element
should or shouldn't be shown when a given theme is active.

This is now used to style the sidebar logo.