Mark up the URL for the GitHub repository
Use an HTML data attribute, data-technote-source-url, to mark up the
HTML element that links to a GitHub or similar repository. This will
allow Ook to scrape the source URL. This isn't a standard, but using a
data attribute like this is valid for internal usage where here
"internal" extends to the Rubin ecosystem.