Add provisional styling for tables
Sphinx adds table elements directly to the HTML. Ideally we actually
want a div around the table so its easier to constrain the table width
and set overflow-x. This approach of display: block work, but it ends up
ruining the accessibility of a table (presumably to deal better with
table layouts). I think the right solution is to add a Sphinx extension
that patches the HTML to add a div.