Render a status aside below h1
The new "insertstatus" sub-extension in technote takes the rendered
document URL and adds an aside with the configured status for the
technote. A technote in a "stable" state does not have a specific status
associated with it.

The HTML for the status is marked up so it's easy to extract data about
the status

- on the aside, the data-technote-status data attribute describes the
  status
- The note is marked up with "technote-status__note"
- Supersceding URLs are marked up with "technote-supersceding-url"