Refactor the original technote.config module
There are now separate classes for modeling `technote.toml` settings,
internal representation of metadata
(`technote.metadata.model.TechnoteMetadata`), the front-end for
bootstrapping the Sphinx configuration in `conf.py`
(`technote.main.TechnoteSphinxConfig`), and the Jinja context for HTML
templating (`technote.templating.context.TechnoteJinjaContext`). A new
factory class is responsible for creating these objects
(`technote.factory.TechnoteFactory`). The benefit of this architecture
is that it's not possible to continue building metadata from additional
sources (external APIs and the technote context) since the metadata is
not longer strictly tied to the `technote.toml` file representation.