Dynamically detect root source file
Previously we assumed the source file was "index.rst" for the purpose of
building the edit-on-github link. Now, in the factory
TechnoteSphinxConfig we check if index.rst, index.md, or index.ipynb
exist, and then set that path as appropriate. This propagates to
TechnoteJinjaContext where the edit on github link is built.