Require given and family names
Remove the "name" key in [[technote.authors.name]] and now require that
authors have a given and family name. Also rename "given_names" to
"given" and "family_names" to "family", since that makes sense in the
context of a PersonName model.

This will allow us to crosswalk terminology to CITATION.cff and other
bibliographic formats.