Allow native toml dates
TOML has native support for date and datetime types (without them being
strings). To accomodate this, the type in TechnoteTable needs to allow
datetime.date types to allow Pydantic to validate those values. The
properties date_update_datetime and date_created_datetime demonstrate to
the typing system that the values are normalized to a datetime.