Change html font size to 100%
Move setting the default font size to the body. This gives us slightly
bigger text while respecting the user stylesheet.

Also make the inline code font slightly smaller to help with baseline
alignment issues for inline monospace text seen in Safari.

Since the rem is now smaller, made the sizes in layout.scss use ems to
be relative to the size set on the body.