Pin pytest < 8 to maintiain isolated test builds
I found that in pytest 8, running multiple builds for different test
roots with the sphinx text mark reused the same import of technote,
which in turn means that the metadata is not being rebuilt. This is fine
for normal sphinx builds where technote's technote.sphinxconf is being
imported only once, and therefore is dedicated to that single sphinx
build.

I think the solution is to make a new project builder for tests that
runs sphinx under a subprocess so that we can inspect the output at the
end.