Standardize and set left padding for block
The goal here is to apply a negative margin to Sphinx block elements
like code blocks and admintions so that their text content lines up with
the body text.

If they have a border or background, then their container needs to have
padding big enough to show this overhang (this is a side-effect of
needing to set the overflow property on the article to support
independent scrolling).

For this to work reliably, I'm creating a new custom property so capture
this dimension so that it can be used reliably and consistently for both
the padding on article and the negative margin on the block elements.